LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.3_Line_Terminators/S7.3_A2.2_T2-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Unexpected EOF
S7.3_A2.2_T2

PASS Expected parsing failure

TEST COMPLETE
